// Assets-Uncapped/front-portal-develop/src/services/api/agreements/models/DrawRepaymentPreviewResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * api-agreements
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: Unknown version
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { MonetaryAmount } from './MonetaryAmount';
import {
    MonetaryAmountFromJSON,
    MonetaryAmountFromJSONTyped,
    MonetaryAmountToJSON,
    MonetaryAmountToJSONTyped,
} from './MonetaryAmount';

/**
 * 
 * @export
 * @interface DrawRepaymentPreviewResponse
 */
export interface DrawRepaymentPreviewResponse {
    /**
     * 
     * @type {Date}
     * @memberof DrawRepaymentPreviewResponse
     */
    repaymentStartDate?: Date;
    /**
     * 
     * @type {number}
     * @memberof DrawRepaymentPreviewResponse
     */
    numberOfRepayments?: number;
    /**
     * 
     * @type {number}
     * @memberof DrawRepaymentPreviewResponse
     */
    drawFeePerMonth?: number;
    /**
     * 
     * @type {MonetaryAmount}
     * @memberof DrawRepaymentPreviewResponse
     */
    totalRepayment?: MonetaryAmount;
    /**
     * 
     * @type {MonetaryAmount}
     * @memberof DrawRepaymentPreviewResponse
     */
    singleRepayment?: MonetaryAmount;
    /**
     * 
     * @type {MonetaryAmount}
     * @memberof DrawRepaymentPreviewResponse
     */
    drawFee?: MonetaryAmount;
    /**
     * 
     * @type {number}
     * @memberof DrawRepaymentPreviewResponse
     */
    drawPercentFee?: number;
}

/**
 * Check if a given object implements the DrawRepaymentPreviewResponse interface.
 */
export function instanceOfDrawRepaymentPreviewResponse(value: object): value is DrawRepaymentPreviewResponse {
    return true;
}

export function DrawRepaymentPreviewResponseFromJSON(json: any): DrawRepaymentPreviewResponse {
    return DrawRepaymentPreviewResponseFromJSONTyped(json, false);
}

export function DrawRepaymentPreviewResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): DrawRepaymentPreviewResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'repaymentStartDate': json['repaymentStartDate'] == null ? undefined : (new Date(json['repaymentStartDate'])),
        'numberOfRepayments': json['numberOfRepayments'] == null ? undefined : json['numberOfRepayments'],
        'drawFeePerMonth': json['drawFeePerMonth'] == null ? undefined : json['drawFeePerMonth'],
        'totalRepayment': json['totalRepayment'] == null ? undefined : MonetaryAmountFromJSON(json['totalRepayment']),
        'singleRepayment': json['singleRepayment'] == null ? undefined : MonetaryAmountFromJSON(json['singleRepayment']),
        'drawFee': json['drawFee'] == null ? undefined : MonetaryAmountFromJSON(json['drawFee']),
        'drawPercentFee': json['drawPercentFee'] == null ? undefined : json['drawPercentFee'],
    };
}

export function DrawRepaymentPreviewResponseToJSON(json: any): DrawRepaymentPreviewResponse {
    return DrawRepaymentPreviewResponseToJSONTyped(json, false);
}

export function DrawRepaymentPreviewResponseToJSONTyped(value?: DrawRepaymentPreviewResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'repaymentStartDate': value['repaymentStartDate'] == null ? undefined : ((value['repaymentStartDate']).toISOString().substring(0,10)),
        'numberOfRepayments': value['numberOfRepayments'],
        'drawFeePerMonth': value['drawFeePerMonth'],
        'totalRepayment': MonetaryAmountToJSON(value['totalRepayment']),
        'singleRepayment': MonetaryAmountToJSON(value['singleRepayment']),
        'drawFee': MonetaryAmountToJSON(value['drawFee']),
        'drawPercentFee': value['drawPercentFee'],
    };
}