// Assets-Uncapped/front-portal-develop/src/services/api/agreements/models/DrawResponse.ts
export interface DrawResponse {
    /**
     * 
     * @type {string}
     * @memberof DrawResponse
     */
    id?: string;
    /**
     * 
     * @type {string}
     * @memberof DrawResponse
     */
    lineOfCreditId?: string;
    /**
     * 
     * @type {MonetaryAmount}
     * @memberof DrawResponse
     */
    size?: MonetaryAmount;
    /**
     * 
     * @type {string}
     * @memberof DrawResponse
     */
    status?: DrawResponseStatusEnum;
    /**
     * 
     * @type {string}
     * @memberof DrawResponse
     */
    name?: string;
    /**
     * 
     * @type {Date}
     * @memberof DrawResponse
     */
    activationDate?: Date;
    /**
     * 
     * @type {string}
     * @memberof DrawResponse
     */
    agreementId?: string;
    /**
     * 
     * @type {string}
     * @memberof DrawResponse
     */
    rejectionReason?: string;
    /**
     * 
     * @type {DrawSigningDetails}
     * @memberof DrawResponse
     */
    drawSigningDetails?: DrawSigningDetails;
    /**
     * 
     * @type {number}
     * @memberof DrawResponse
     */
    repaymentTermsMonths?: number;
    /**
     * 
     * @type {number}
     * @memberof DrawResponse
     */
    numberOfRepayments?: number;
    /**
     * 
     * @type {number}
     * @memberof DrawResponse
     */
    totalPercentFee?: number;
    /**
     * 
     * @type {string}
     * @memberof DrawResponse
     */
    cashTransferAccountId?: string;
    /**
     * 
     * @type {boolean}
     * @memberof DrawResponse
     */
    isFirstDraw?: boolean;
    /**
     * 
     * @type {RepaymentCustomisationParametersResponse}
     * @memberof DrawResponse
     */
    repaymentCustomisationParametersResponse?: RepaymentCustomisationParametersResponse;
    /**
     * 
     * @type {string}
     * @memberof DrawResponse
     */
    externalId?: string;
    /**
     * 
     * @type {Set<string>}
     * @memberof DrawResponse
     */
    documentIds?: Set<string>;
    /**
     * 
     * @type {Date}
     * @memberof DrawResponse
     */
    requestedAtDate?: Date;
    /**
     * 
     * @type {string}
     * @memberof DrawResponse
     */
    dealId?: string;
}


/**
 * @export
 */
export const DrawResponseStatusEnum = {
    Draft: 'DRAFT',
    WaitingForSignature: 'WAITING_FOR_SIGNATURE',
    Pending: 'PENDING',
    Active: 'ACTIVE',
    Rejected: 'REJECTED',
    Closed: 'CLOSED',
    Unknown: 'UNKNOWN'
} as const;
export type DrawResponseStatusEnum = typeof DrawResponseStatusEnum[keyof typeof DrawResponseStatusEnum];