// Assets-Uncapped/front-portal-develop/src/services/api/agreements/models/LineOfCreditResponse.ts
export interface LineOfCreditResponse {
    /**
     * 
     * @type {string}
     * @memberof LineOfCreditResponse
     */
    id?: string;
    /**
     * 
     * @type {string}
     * @memberof LineOfCreditResponse
     */
    organisationId?: string;
    /**
     * 
     * @type {string}
     * @memberof LineOfCreditResponse
     */
    name?: string;
    /**
     * 
     * @type {MonetaryAmount}
     * @memberof LineOfCreditResponse
     */
    limit?: MonetaryAmount;
    /**
     * 
     * @type {LineOfCreditBalance}
     * @memberof LineOfCreditResponse
     */
    balance?: LineOfCreditBalance;
    /**
     * 
     * @type {string}
     * @memberof LineOfCreditResponse
     */
    status?: LineOfCreditResponseStatusEnum;
    /**
     * 
     * @type {Date}
     * @memberof LineOfCreditResponse
     */
    creationDate?: Date;
    /**
     * 
     * @type {Date}
     * @memberof LineOfCreditResponse
     */
    lastUnderwritingDate?: Date;
    /**
     * 
     * @type {Date}
     * @memberof LineOfCreditResponse
     */
    drawDownPeriodEndDate?: Date;
    /**
     * 
     * @type {FeesParametersResponse}
     * @memberof LineOfCreditResponse
     */
    fees?: FeesParametersResponse;
    /**
     * 
     * @type {DrawParametersResponse}
     * @memberof LineOfCreditResponse
     */
    drawParameters?: DrawParametersResponse;
    /**
     * 
     * @type {string}
     * @memberof LineOfCreditResponse
     */
    collectionFrequency?: LineOfCreditResponseCollectionFrequencyEnum;
    /**
     * 
     * @type {string}
     * @memberof LineOfCreditResponse
     */
    repaymentType?: LineOfCreditResponseRepaymentTypeEnum;
    /**
     * 
     * @type {DeferredRepaymentsResponse}
     * @memberof LineOfCreditResponse
     */
    deferredRepayments?: DeferredRepaymentsResponse;
    /**
     * 
     * @type {Array<DrawResponse>}
     * @memberof LineOfCreditResponse
     */
    draws?: Array<DrawResponse>;
    /**
     * 
     * @type {string}
     * @memberof LineOfCreditResponse
     */
    dealId?: string;
    /**
     * 
     * @type {string}
     * @memberof LineOfCreditResponse
     */
    type?: LineOfCreditResponseTypeEnum;
    /**
     * 
     * @type {AgreementId}
     * @memberof LineOfCreditResponse
     */
    agreementId?: AgreementId;
    /**
     * 
     * @type {RepaymentTerms}
     * @memberof LineOfCreditResponse
     */
    repaymentTerms?: RepaymentTerms;
    /**
     * 
     * @type {number}
     * @memberof LineOfCreditResponse
     */
    availableCreditLimitMultiplier?: number;
    /**
     * 
     * @type {number}
     * @memberof LineOfCreditResponse
     */
    maxCreditLimit?: number;
    /**
     * 
     * @type {number}
     * @memberof LineOfCreditResponse
     */
    currentAvailableCreditLimit?: number;
}


/**
 * @export
 */
export const LineOfCreditResponseStatusEnum = {
    Active: 'ACTIVE',
    Paused: 'PAUSED',
    WaitingForDocuments: 'WAITING_FOR_DOCUMENTS',
    WaitingForApproval: 'WAITING_FOR_APPROVAL',
    DrawPeriodEnded: 'DRAW_PERIOD_ENDED',
    PendingClosure: 'PENDING_CLOSURE',
    Closed: 'CLOSED',
    Unknown: 'UNKNOWN'
} as const;
export type LineOfCreditResponseStatusEnum = typeof LineOfCreditResponseStatusEnum[keyof typeof LineOfCreditResponseStatusEnum];

/**
 * @export
 */
export const LineOfCreditResponseCollectionFrequencyEnum = {
    Daily: 'DAILY',
    Weekly: 'WEEKLY',
    Every14Days: 'EVERY_14_DAYS',
    Every15Days: 'EVERY_15_DAYS',
    Monthly: 'MONTHLY',
    OnDemand: 'ON_DEMAND',
    Unknown: 'UNKNOWN'
} as const;
export type LineOfCreditResponseCollectionFrequencyEnum = typeof LineOfCreditResponseCollectionFrequencyEnum[keyof typeof LineOfCreditResponseCollectionFrequencyEnum];

/**
 * @export
 */
export const LineOfCreditResponseRepaymentTypeEnum = {
    Automatic: 'AUTOMATIC',
    Manual: 'MANUAL',
    Unknown: 'UNKNOWN'
} as const;
export type LineOfCreditResponseRepaymentTypeEnum = typeof LineOfCreditResponseRepaymentTypeEnum[keyof typeof LineOfCreditResponseRepaymentTypeEnum];

/**
 * @export
 */
export const LineOfCreditResponseTypeEnum = {
    InterestRate: 'INTEREST_RATE',
    Fixed: 'FIXED',
    Unknown: 'UNKNOWN'
} as const;
export type LineOfCreditResponseTypeEnum = typeof LineOfCreditResponseTypeEnum[keyof typeof LineOfCreditResponseTypeEnum];